pyproject.toml
[project]
name = "autograd-ml"
version = "0.1.0"
description = "Training model pipeline"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "catboost>=1.2.8",
    "dvc[s3]>=3.65.0",
    "hydra-core>=1.3.2",
    "mlflow>=3.8.1",
    "pre-commit>=4.5.1",
    "sentence-transformers>=5.2.0",
]

[tool.yapf]
column_limit = 88
based_on_style = "PEP8"
INDENT_DICTIONARY_VALUE = true

[tool.ruff.lint]
ignore = ["E701", "E731"]
